requirements.txt
# Core dependencies
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.11.1

# Data processing
yfinance==0.2.18
ccxt==4.0.77
requests==2.31.0
websocket-client==1.6.3

# Technical analysis
ta-lib==0.4.25
talib==0.4.25

# Machine Learning
xgboost==1.7.6
lightgbm==4.0.0
tensorflow==2.13.0
torch==2.0.1

# Web framework
streamlit==1.25.0
plotly==5.15.0
dash==2.11.1
dash-bootstrap-components==1.4.1

# Database
sqlite3
sqlalchemy==2.0.19
alembic==1.11.1

# Telegram
python-telegram-bot==20.3

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Logging
loguru==0.7.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Development
black==23.7.0
flake8==6.0.0
mypy==1.5.1

# Deployment
gunicorn==21.2.0
docker==6.1.3
